+++ CHANGELOG.md
@@ -69,7 +69,7 @@
 
 ### 🧰 Bug fixes 🧰
 
-- `signalfexporter`: add no lint error checks for signalf exporter package (#10258)
+- `signalfxexporter`: add no lint error checks for signalfx exporter package (#10258)
 - `k8sclusterreceiver`: Fix the receiver to work with 1.19 and 1.20 k8s API versions (#9523)
 - `azuremonitorexporter`: Fix log exporter bug related to incorrectly mapping SpanId (#9579)
 - `mysqlreceiver`: Fix attribute values mismatch with its definition (#9688)